Move code samples into markdown cells
My understanding is that these code snippets are not meant to be
executed. Moving them into markdown cells makes it harder for the user
to mistakenly run the cells, since doing so would add an unwanted extra
Dense layer to the user's model.

The result of running the `model.add(..)` cell is that when the user
later pops the last layer for fine-tuning, they pop this extraneous
Dense layer rather than the original Vgg16 output layer. If the user
later saves the model weights and attempts to load them into a fresh
model in lesson3, they'll get an exception about trying to load a
weights file with 17 layers into a model with 16 layers.

Possibly related forum thread:

http://forums.fast.ai/t/lesson-3-discussion/186/35