Use np.squeeze and np.atleast_3d to munge dimensions for {x,y}_rnn
If you run the cells of this notebook sequentially, then you'll get
one of the following exceptions when attempting to call `model.fit` in
the "Sequence model with keras" section:

    Exception: Error when checking model input: expected
    embedding_input_2 to have 2 dimensions, but got array with
    shape (75110, 8, 1)

    Exception: Error when checking model target: expected
    timedistributed_1 to have 3 dimensions, but got array with
    shape (75110, 8, 1, 1)

The problem is that `xs` and `ys` will have different dimensions
depending on whether you ran all previous cells of the notebook, or
whether you ran only the minimum cells required to work on the current
section.

The solution is to modify `x_rnn` and `y_rnn` so that they have the
correct shape regardless of whether all previous cells have been run
or only the cells required to define `xs` and `ys`.